requirements.txt
selenium
pytest>=7.0.0
webdriver-manager==3.5.0
pillow>=10.0.0
allure-pytest
eyes-selenium
python-dotenv



